+++ frontend/.angular/cache/17.3.10/vite/deps/@angular_router.js
@@ -1,13 +1,15 @@
 import {
+  Title
+} from "./chunk-QIQS55LB.js";
+import {
   DOCUMENT,
   HashLocationStrategy,
   LOCATION_INITIALIZED,
   Location,
   LocationStrategy,
   PathLocationStrategy,
-  Title,
   ViewportScroller
-} from "./chunk-WQ3DZVEW.js";
+} from "./chunk-A3LTE4ZH.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   APP_INITIALIZER,
