+++ frontend/.angular/cache/17.3.10/vite/deps/WebGPURenderer-IO57YZTQ.js
@@ -1,11 +1,11 @@
 import {
   WebGPURenderer
-} from "./chunk-Q2H3BWCN.js";
-import "./chunk-REVMUJ3R.js";
+} from "./chunk-EMKP3YEF.js";
 import "./chunk-ZT4WG52D.js";
+import "./chunk-REVMUJ3R.js";
 import "./chunk-2ABCISNK.js";
 import "./chunk-ASLTLD6L.js";
 export {
   WebGPURenderer
 };
-//# sourceMappingURL=WebGPURenderer-3H25LEES.js.map
+//# sourceMappingURL=WebGPURenderer-IO57YZTQ.js.map
