+++ frontend/.angular/cache/17.3.10/vite/deps/_metadata.json
@@ -1,37 +1,52 @@
 {
-  "hash": "1923ff01",
-  "configHash": "2ef6d8d8",
+  "hash": "766541dc",
+  "configHash": "407886e8",
   "lockfileHash": "609c37b7",
-  "browserHash": "e115956b",
+  "browserHash": "d23760e6",
   "optimized": {
     "@angular/core": {
       "src": "../../../../../node_modules/@angular/core/fesm2022/core.mjs",
       "file": "@angular_core.js",
-      "fileHash": "d3e59869",
+      "fileHash": "05b90525",
       "needsInterop": false
     },
     "@angular/platform-browser": {
       "src": "../../../../../node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs",
       "file": "@angular_platform-browser.js",
-      "fileHash": "c402c555",
+      "fileHash": "1fb0f503",
       "needsInterop": false
     },
     "@angular/router": {
       "src": "../../../../../node_modules/@angular/router/fesm2022/router.mjs",
       "file": "@angular_router.js",
-      "fileHash": "dfade964",
+      "fileHash": "eb077c8a",
       "needsInterop": false
     },
     "pixi.js": {
       "src": "../../../../../node_modules/pixi.js/lib/index.mjs",
       "file": "pixi__js.js",
-      "fileHash": "f4d58a29",
+      "fileHash": "db62a2e6",
+      "needsInterop": false
+    },
+    "@angular/common/http": {
+      "src": "../../../../../node_modules/@angular/common/fesm2022/http.mjs",
+      "file": "@angular_common_http.js",
+      "fileHash": "342fe648",
       "needsInterop": false
     }
   },
   "chunks": {
-    "chunk-WQ3DZVEW": {
-      "file": "chunk-WQ3DZVEW.js"
+    "WebGLRenderer-XT5ELZWL": {
+      "file": "WebGLRenderer-XT5ELZWL.js"
+    },
+    "chunk-3JWMVLXI": {
+      "file": "chunk-3JWMVLXI.js"
+    },
+    "chunk-QIQS55LB": {
+      "file": "chunk-QIQS55LB.js"
+    },
+    "chunk-A3LTE4ZH": {
+      "file": "chunk-A3LTE4ZH.js"
     },
     "chunk-LGGNAV7W": {
       "file": "chunk-LGGNAV7W.js"
@@ -51,23 +66,17 @@
     "chunk-XYGGB2JN": {
       "file": "chunk-XYGGB2JN.js"
     },
-    "WebGPURenderer-3H25LEES": {
-      "file": "WebGPURenderer-3H25LEES.js"
+    "WebGPURenderer-IO57YZTQ": {
+      "file": "WebGPURenderer-IO57YZTQ.js"
     },
-    "chunk-Q2H3BWCN": {
-      "file": "chunk-Q2H3BWCN.js"
-    },
-    "chunk-REVMUJ3R": {
-      "file": "chunk-REVMUJ3R.js"
-    },
-    "WebGLRenderer-XT5ELZWL": {
-      "file": "WebGLRenderer-XT5ELZWL.js"
-    },
-    "chunk-3JWMVLXI": {
-      "file": "chunk-3JWMVLXI.js"
+    "chunk-EMKP3YEF": {
+      "file": "chunk-EMKP3YEF.js"
     },
     "chunk-ZT4WG52D": {
       "file": "chunk-ZT4WG52D.js"
+    },
+    "chunk-REVMUJ3R": {
+      "file": "chunk-REVMUJ3R.js"
     },
     "chunk-2ABCISNK": {
       "file": "chunk-2ABCISNK.js"
